fix: mutualise ObjectMapper configuration
