+++ scribe-core/src/main/java/fr/maif/json/MapperSingleton.java
@@ -1,0 +1,26 @@
+package fr.maif.json;
+
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+
+public final class MapperSingleton {
+    public final static ObjectMapper mapper = buildMapper();
+
+    // Prevent instantiation
+    private MapperSingleton() { }
+
+    public static ObjectMapper getInstance() {
+        return mapper;
+    }
+
+    private static ObjectMapper buildMapper() {
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.registerModule(new JavaTimeModule());
+        mapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false);
+        mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+        mapper.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false);
+        return mapper;
+    }
+}
